+++ addon/utils/types.ts
@@ -1,0 +1,3 @@
+import { htmlSafe } from '@ember/template';
+
+export type SafeString = ReturnType<typeof htmlSafe>;