+++ app/components/fragment-plugin/fragments/fragment-list.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-rdfa-editor-lblod-plugins/components/fragment-plugin/fragments/fragment-list';